EV-103 Add Routing, fix db connection, add schemas to some reponses and requests
