+++ Dockerfile
@@ -8,4 +8,4 @@ RUN pip install --no-cache-dir --upgrade -r ./code/requirements.txt
 
 COPY ./src /code/src
 
-CMD [ "uvicorn", "src.main:app", "--host", "0.0.0.0", "--port", "80" ]
+CMD [ "uvicorn", "src.server.app:app", "--host", "0.0.0.0", "--port", "80" ]
